requirements.txt
Django==2.1.2
django-appconf==1.0.2
django-crispy-forms==1.7.2
django-filter==2.0.0
django-imagekit==4.0.2
pilkit==2.0
Pillow==5.3.0
pytz==2018.5
six==1.11.0
